# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Desk Monitor Platform API
  version: 1.0.0
  description: Contract-first API for the administration console, devices and event producers.
servers:
  - url: /api/v1
...
paths:
...
  /services:
    get:
      tags: [Services]
      operationId: listServices
      security: [{userBearer: []}]
      parameters:
        - $ref: '#/components/parameters/ProductFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Service page.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/Service'}}
    post:
      tags: [Services]
      operationId: createService
      security: [{userBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ServiceCreate'}
      responses:
        '201':
          description: Service created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Service'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /services/{serviceId}:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
    get:
      tags: [Services]
      operationId: getService
      security: [{userBearer: []}]
      responses:
        '200':
          description: Service detail.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Service'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Services]
      operationId: updateService
      security: [{userBearer: []}]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: {$ref: '#/components/schemas/ServiceUpdate'}
      responses:
        '200':
          description: Service updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Service'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Services]
      operationId: archiveService
      security: [{userBearer: []}]
      responses:
        '204': {description: Service archived.}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /alerts:
    get:
      tags: [Alerts]
      operationId: listAlerts
      security: [{userBearer: []}]
      parameters:
        - {name: state, in: query, schema: {$ref: '#/components/schemas/AlertState'}}
        - {name: priority, in: query, schema: {$ref: '#/components/schemas/Priority'}}
        - $ref: '#/components/parameters/ProductFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Alert page.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/Alert'}}
    post:
      tags: [Alerts]
      operationId: createManualAlert
      security: [{userBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AlertCreate'}
      responses:
        '201':
          description: Manual alert created and queued for publication.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Alert'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /alerts/{alertId}/acknowledge:
    post:
      tags: [Alerts]
      operationId: acknowledgeAlert
      security: [{userBearer: []}]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      responses:
        '200':
          description: Alert acknowledged idempotently.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Alert'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /alerts/{alertId}/resolve:
    post:
      tags: [Alerts]
      operationId: resolveAlert
      security: [{userBearer: []}]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      responses:
        '200':
          description: Alert resolved idempotently.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Alert'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    userBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    deviceBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    githubActionsBearer: {type: http, scheme: bearer}
    uptimeBearer: {type: http, scheme: bearer}
    refreshCookie: {type: apiKey, in: cookie, name: desk_refresh}
  parameters:
...
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
    ProductFilter: {name: product_id, in: query, schema: {type: string}}
...
    ServiceId: {name: serviceId, in: path, required: true, schema: {type: string}}
    AlertId: {name: alertId, in: path, required: true, schema: {type: string}}
...
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    NotFound:
      description: Resource not found.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Conflict:
      description: State or uniqueness conflict.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, request_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        request_id: {type: string}
        errors: {type: object, additionalProperties: {type: array, items: {type: string}}}
...
    Page:
      type: object
      required: [items, has_more]
      properties:
        items: {type: array, items: {}}
        next_cursor: {type: [string, 'null']}
        has_more: {type: boolean}
...
    Service:
      type: object
      required: [id, product_id, key, name, status, updated_at]
      properties:
        id: {type: string}
        product_id: {type: string}
        key: {type: string}
        name: {type: string}
        status: {$ref: '#/components/schemas/ServiceState'}
        latency_ms: {type: [integer, 'null'], minimum: 0}
        updated_at: {type: string, format: date-time}
    ServiceCreate:
      type: object
      additionalProperties: false
      required: [product_id, key, name]
      properties:
        product_id: {type: string}
        key: {type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}$'}
        name: {type: string, minLength: 1, maxLength: 100}
        kind: {type: string, maxLength: 64}
        endpoint: {type: [string, 'null'], format: uri}
        enabled: {type: boolean, default: true}
    ServiceUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        kind: {type: string, maxLength: 64}
        endpoint: {type: [string, 'null'], format: uri}
        enabled: {type: boolean}
    ServiceState: {type: string, enum: [operational, degraded, outage, maintenance, unknown]}
    Priority: {type: string, enum: [critical, warning, info, success]}
    AlertState: {type: string, enum: [open, acknowledged, resolved]}
    Alert:
      type: object
      required: [id, fingerprint, priority, state, title, message, opened_at, updated_at]
      properties:
        id: {type: string}
        product_id: {type: [string, 'null']}
        service_id: {type: [string, 'null']}
        fingerprint: {type: string}
        priority: {$ref: '#/components/schemas/Priority'}
        state: {$ref: '#/components/schemas/AlertState'}
        title: {type: string}
        message: {type: string}
        opened_at: {type: string, format: date-time}
        acknowledged_at: {type: [string, 'null'], format: date-time}
        resolved_at: {type: [string, 'null'], format: date-time}
        updated_at: {type: string, format: date-time}
    AlertCreate:
      type: object
      additionalProperties: false
      required: [fingerprint, priority, title, message]
      properties:
        product_id: {type: [string, 'null']}
        service_id: {type: [string, 'null']}
        fingerprint: {type: string, minLength: 1, maxLength: 160}
        priority: {$ref: '#/components/schemas/Priority'}
        title: {type: string, minLength: 1, maxLength: 48}
        message: {type: string, minLength: 1, maxLength: 160}
        expires_at: {type: [string, 'null'], format: date-time}
...